requirements.txt
Django~=5.1.3
djangorestframework~=3.15.2
drf-yasg~=1.21.8
djangorestframework-simplejwt~=5.3.1
django-jazzmin
django-cors-headers